Adapted the code for Python 3

+++ tests/test_continuous_injection.py
@@ -103,7 +103,7 @@ def run_continuous_injection( gamma_boost, dens_func,
     sim.set_moving_window( v=c, gamma_boost=gamma_boost )
 
     # Check that the density is correct after different timesteps
-    N_step = Nz/N_check/2
+    N_step = int( Nz/N_check/2 )
     for i in range( N_check ):
         sim.step( N_step, move_momenta=False )
         check_density( sim, gamma_boost, dens_func, show )
